ChatMessageの切り出し
