+++ churaverse-plugins-client/src/meetingPlugins/meetingCoreUiPlugin/domain/chatMessage.ts
@@ -1,0 +1,9 @@
+export interface ChatMessage {
+  type: 'chat' | 'chat_history'
+  messageId?: string
+  sender: string
+  senderId?: string
+  text: string
+  timestamp: number
+  history?: ChatMessage[]
+}
