+++ churaverse-plugins-client/src/meetingPlugins/meetingCoreUiPlugin/meetingWebRtcPlugin.ts
@@ -17,8 +17,9 @@ import {
 } from './components/MeetingControlBarComponent'
 import { VIDEO_GRID_ID } from './components/VideoGridComponent'
 import { CHAT_INPUT_ID, CHAT_SEND_BUTTON_ID } from './components/MeetingSidebarComponent'
-import { MeetingPluginStore, ChatMessage } from './store/defMeetingPluginStore'
+import { MeetingPluginStore } from './store/defMeetingPluginStore'
 import { initMeetingPluginStore } from './store/initMeetingPluginStore'
+import { ChatMessage } from './domain/chatMessage'
 import { VideoGridUi } from './ui/videoGridUi'
 import { ParticipantListUi } from './ui/participantListUi'
 import { ChatUi } from './ui/chatUi'
